feat : [BE] MVP 베팅 스마트컨트랙트 함수 호출 [S12P21A609-471]

+++ src/main/java/org/example/betty/external/game/scraper/LineupScraper.java
@@ -90,7 +90,7 @@ public class LineupScraper extends BaseScraper {
         List<WebElement> playerItems = teamElement.findElements(By.cssSelector(".Lineup_lineup_item__32s4M"));
         List<PlayerInfo> players = new ArrayList<>();
 
-        int playerIdSeq = playerIdStart;
+        long playerIdSeq = playerIdStart;
 
         for (WebElement item : playerItems) {
             try {
